requirements.txt
aiogram~=3.18.0
aiogram_dialog~=2.3.1
autoscraper
pandas
psycopg
plotly
pydantic~=2.10.6
pydantic-settings~=2.8.1
Pyrogram~=2.0.106
python-dateutil
python-dotenv
python-json-logger
pytz~=2024.2
regex
requests
scipy~=1.14.1
setuptools
SQLAlchemy~=2.0.36
voyageai
fuzzywuzzy
matplotlib
spacy
redis